+++ docs/js/general.js
@@ -37,7 +37,7 @@ function create() {
 function create_newboard() {
 
     var size = parseInt(document.getElementById("nb_size3").value);
-    if (15 <= size && size <= 80) {
+    if (12 <= size && size <= 90) {
         var mode = pu.mode;
         var gridtype = document.getElementById("gridtype").value;
         pu = make_class(gridtype);
@@ -48,7 +48,7 @@ function create_newboard() {
         document.getElementById('modal').style.display = 'none';
         pu.mode_set(pu.mode[pu.mode.qa].edit_mode); //include redraw
     } else {
-        alert("Display size must be in the range 15-80");
+        alert("Display size must be in the range 12-90");
     }
 }
 
@@ -276,27 +276,27 @@ function CreateCheck() {
 
 function newgrid() {
     var size = parseInt(document.getElementById("nb_size3").value);
-    if (15 <= size && size <= 80) {
+    if (12 <= size && size <= 90) {
         pu.reset_frame_newgrid();
         pu.redraw();
         panel_pu.draw_panel();
         document.getElementById('modal').style.display = 'none';
     } else {
-        alert("Display size must be in the range 15-80");
+        alert("Display size must be in the range 12-90");
     }
 }
 
 function newgrid_r() {
     var sizer = parseInt(document.getElementById("nb_size3_r").value, 10);
     document.getElementById("nb_size3").value = sizer;
-    if (15 <= sizer && sizer <= 80) {
+    if (12 <= sizer && sizer <= 90) {
         pu.reset_frame_newgrid();
         pu.size = sizer;
         pu.redraw();
         panel_pu.draw_panel();
         document.getElementById('modal-newsize').style.display = 'none';
     } else {
-        alert("Display size must be in the range 15-80");
+        alert("Display size must be in the range 12-90");
     }
 }